I realized everything I had in item.py belonged in manager.py
